CRUD da entidade Autor implementado

+++ backend/src/main/java/br/com/videolocadorapassatempo/service/impl/ActorServiceImpl.java
@@ -4,14 +4,13 @@ import br.com.videolocadorapassatempo.model.ActorModel;
 import br.com.videolocadorapassatempo.repository.ActorRepository;
 import br.com.videolocadorapassatempo.service.ActorService;
 import br.com.videolocadorapassatempo.service.dto.ActorDto;
-import br.com.videolocadorapassatempo.service.exception.ActorException;
+import br.com.videolocadorapassatempo.service.exception.EntityException;
 import br.com.videolocadorapassatempo.service.mapper.ActorMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -27,21 +26,37 @@ public class ActorServiceImpl implements ActorService {
     }
 
     public ActorDto findById(Long id) {
-        Optional<ActorModel> actorModel = actorRepository.findById(id);
+        log.info("Buscando ator de id = {}", id);
+        ActorModel actorModel = actorRepository.findById(id)
+                .orElseThrow(() -> new EntityException("Autor de id = " + id + " não encontrado!"));
 
-        if(!actorModel.isPresent()) {
-            throw new ActorException("Autor não encontrado");
-        }
-
-        return actorMapper.toDto(actorModel.get());
+        log.info("Ator de id = {} encontrado com sucesso", id);
+        return actorMapper.toDto(actorModel);
     }
 
-    public ActorDto save(ActorDto actorDto) {
+    public ActorDto create(ActorDto actorDto) {
+        log.info("Ator cadastrado com sucesso!");
         ActorModel actorModel = actorMapper.toEntity(actorDto);
         return actorMapper.toDto(actorRepository.save(actorModel));
     }
 
+    public ActorDto update(ActorDto actorDto) {
+        log.info("Buscando ator de id = {}", actorDto.getId());
+        ActorModel actorModel = actorRepository.findById(actorDto.getId())
+                .orElseThrow(() -> new EntityException("Autor de id = " + actorDto.getId() + " não encontrado!"));
+
+        actorModel.setName(actorDto.getName());
+        log.info("Ator de id = {} atualizado com sucesso!", actorDto.getId());
+        return actorMapper.toDto(actorRepository.save(actorModel));
+    }
+
     public void deleteById(Long id) {
+        log.info("Buscando ator de id = {}", id);
+        if(!actorRepository.existsById(id)) {
+            throw new EntityException("Autor com id = " + id + " não encontrado!");
+        }
+
+        log.info("Ator de id = {} deletado com sucesso!", id);
         actorRepository.deleteById(id);
     }
